students/K33402/Velts Andrey/lab0304/backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Django==3.1.4
django-cors-headers==3.5.0
django-hosts==4.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
drf-yasg==1.20.0
easy-thumbnails==2.7.1
flake8==3.8.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.7
pathspec==0.8.1
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pyparsing==2.4.7
pytz==2020.4
regex==2020.11.13
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
